requirements.txt
crewai==0.11.0
langchain==0.1.0
langchain-core==0.1.19
langchain-community==0.0.18
langchain-openai==0.0.2.post1
langsmith==0.0.92
python-dotenv
linkup-sdk
pydantic
streamlit
mcp